src/requirements.txt
objectpath
PyJWT
pytz
requests
pandas
scipy
s3fs
